move whitenoise rpc Server to new file
